Remove glbinding-aux dependency in release mode
