+++ src/backend/opengl/common.hpp
@@ -154,6 +154,10 @@ namespace forge
 namespace opengl
 {
 
+void glErrorCheck(const char *pMsg, const char* pFile, int pLine);
+
+#define CheckGL(msg) glErrorCheck(msg, __FILE__, __LINE__)
+
 typedef unsigned int    uint;
 typedef unsigned short  ushort;
 typedef unsigned char   uchar;
