+++ src/backend/opengl/font_atlas_impl.cpp
@@ -53,7 +53,6 @@
 
 #include <font_atlas_impl.hpp>
 #include <common.hpp>
-#include <err_opengl.hpp>
 
 #include <cstring>
 #include <algorithm>
